G2-1747 Exit codes are bytes

+++ ax-command/src/test/java/com/g2forge/alexandria/command/invocation/runner/TestICommandRunner.java
@@ -125,7 +125,7 @@ public class TestICommandRunner {
 		}
 
 		final HCLIReport.Output expected = HCLIReport.computeExpectedOutput(invocation.getArguments());
-		HAssert.assertEquals(expected.getExitCode(), exitCode);
+		HAssert.assertEquals((byte) expected.getExitCode(), (byte) exitCode);
 		HAssert.assertEquals(expected.getOutput(), output);
 	}
 }
